strip traceparents in attributes with driver adapters
