+++ query-engine/driver-adapters/Cargo.toml
@@ -15,6 +15,7 @@ once_cell = "1.15"
 prisma-metrics.path = "../../libs/metrics"
 serde.workspace = true
 serde_json.workspace = true
+telemetry.path = "../../libs/telemetry"
 tracing.workspace = true
 tracing-core = "0.1"
 uuid.workspace = true
